+++ goldenverba/components/managers.py
@@ -24,6 +24,8 @@ from goldenverba.components.embedding.CohereEmbedder import CohereEmbedder
 from goldenverba.components.embedding.MiniLMEmbedder import MiniLMEmbedder
 from goldenverba.components.embedding.GoogleEmbedder import GoogleEmbedder
 from goldenverba.components.embedding.OllamaEmbedder import OllamaEmbedder
+from goldenverba.components.embedding.AllMPNetEmbedder import AllMPNetEmbedder
+from goldenverba.components.embedding.MixedbreadEmbedder import MixedbreadEmbedder
 
 from goldenverba.components.retriever.WindowRetriever import WindowRetriever
 
@@ -169,6 +171,8 @@ class EmbeddingManager:
         self.embedders: dict[str, Embedder] = {
             "GoogleEmbedder": GoogleEmbedder(),
             "MiniLMEmbedder": MiniLMEmbedder(),
+            "AllMPNetEmbedder": AllMPNetEmbedder(),
+            "MixedbreadEmbedder": MixedbreadEmbedder(),
             "ADAEmbedder": ADAEmbedder(),
             "CohereEmbedder": CohereEmbedder(),
             "OllamaEmbedder": OllamaEmbedder(),