+++ goldenverba/verba_manager.py
@@ -58,7 +58,7 @@ class VerbaManager:
     def import_data(
         self, fileData: list[FileData], textValues: list[str], logging: list[dict]
     ) -> list[Document]:
-
+        
         loaded_documents, logging = self.reader_manager.load(
             fileData, textValues, logging
         )
@@ -243,6 +243,21 @@ class VerbaManager:
             self.installed_libraries["pypdf"] = False
 
         try:
+            import docx
+            
+            self.installed_libraries["docx"] = True
+        except Exception:
+            self.installed_libraries["docx"] = False
+            
+        try:
+            import sentence_transformers
+            
+            self.installed_libraries["sentence-transformers"] = True
+        except Exception:
+            self.installed_libraries["sentence-transformers"] = False
+
+
+        try:
             import tiktoken
 
             self.installed_libraries["tiktoken"] = True
@@ -307,6 +322,11 @@ class VerbaManager:
         else:
             self.environment_variables["OLLAMA_MODEL"] = False
 
+        if os.environ.get("OLLAMA_EMBED_MODEL", "") != "":
+            self.environment_variables["OLLAMA_EMBED_MODEL"] = True
+        else:
+            self.environment_variables["OLLAMA_EMBED_MODEL"] = False
+
         # OpenAI API Key
         if os.environ.get("OPENAI_API_KEY", "") != "":
             self.environment_variables["OPENAI_API_KEY"] = True
@@ -329,6 +349,12 @@ class VerbaManager:
             self.environment_variables["GITHUB_TOKEN"] = True
         else:
             self.environment_variables["GITHUB_TOKEN"] = False
+
+        # GitLab Token Key
+        if os.environ.get("GITLAB_TOKEN", "") != "":
+            self.environment_variables["GITLAB_TOKEN"] = True
+        else:
+            self.environment_variables["GITLAB_TOKEN"] = False
 
         # Unstructured Token Key
         if os.environ.get("UNSTRUCTURED_API_KEY", "") != "":
@@ -758,11 +784,14 @@ class VerbaManager:
             .do()
         )
 
-        if results["data"]["Get"][class_name]:
-            msg.warn(f"{document.name} already exists")
-            return True
+        if "data" in results:
+            if results["data"]["Get"][class_name]:
+                msg.warn(f"{document.name} already exists")
+                return True
         else:
-            return False
+            msg.warn(f"Error occured while checking for duplicates: {results}")
+        
+        return False
 
     def check_verba_component(self, component: VerbaComponent) -> tuple[bool, str]:
         return component.check_available(