+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Cancel Generation Button
 - Added Documentation for JSON Files
+- Added GitLabReader (https://github.com/weaviate/Verba/pull/151)
 
 ## Fixed